SYNCOPE-1545: address review
